I23 index 2

+++ app/Http/Controllers/Admin/Marketing/PaginaInicial/CarrosselController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers\Admin\Marketing\PaginaInicial;
 
+use App\Enums\Marketing\PaginaInicial\StatusVersao;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Admin\Marketing\PaginaInicial\Carrossel\AdicionarImagemRequest;
 use App\Models\Marketing\PaginaInicial\Carrossel\CarrosselItem;
@@ -21,11 +22,19 @@ class CarrosselController extends Controller
 
     public function adicionar(Versao $versao): mixed
     {
+        if ($versao->status !== StatusVersao::Criado) {
+            $nome = $versao->status->name;
+            abort(403, "Não é possivel editar uma versao com status $nome");
+        }
         return Inertia::Render('Admin/Marketing/PaginaInicial/CarrosselPrincipal/AdicionarImagem', compact('versao'));
     }
 
     public function salvar(AdicionarImagemRequest $request, Versao $versao): mixed
     {
+        if ($versao->status !== StatusVersao::Criado) {
+            $nome = $versao->status->name;
+            abort(403, "Não é possivel editar uma versao com status $nome");
+        }
         $imagemDesktop = $request->file('imagem_desktop');
         $imagemDesktop->store(config('equipamentos.path_imagens'));
 
@@ -36,6 +45,7 @@ class CarrosselController extends Controller
         $item->versao_id = $versao->id;
         $item->nome_arquivo_desktop = $imagemDesktop->hashName();
         $item->nome_arquivo_mobile = $imagemMobile->hashName();
+        $item->ordem = $versao->carrosselItens()->count() + 1;
         $item->save();
 
         return redirect()->route('admin.marketing.paginaInicial.layout.carrossel.visualizar', $versao->id);
@@ -43,10 +53,22 @@ class CarrosselController extends Controller
 
     public function excluir(Versao $versao, CarrosselItem $item): mixed
     {
+        if ($versao->status !== StatusVersao::Criado) {
+            $nome = $versao->status->name;
+            abort(403, "Não é possivel editar uma versao com status $nome");
+        }
         Storage::delete(config('equipamentos.path_imagens') . '/' . $item->nome_arquivo_desktop);
         Storage::delete(config('equipamentos.path_imagens') . '/' . $item->nome_arquivo_mobile);
         $item->delete();
 
-        return redirect()->route('admin.marketing.paginaInicial.layout', $versao->id);
+        return redirect()->route('admin.marketing.paginaInicial.layout.carrossel.visualizar', $versao->id);
+    }
+
+    public function visualizarItem(Versao $versao, CarrosselItem $item): mixed
+    {
+        return Inertia::Render(
+            'Admin/Marketing/PaginaInicial/CarrosselPrincipal/VisualizarItem',
+            compact('versao', 'item')
+        );
     }
 }
