+++ app/Http/Controllers/Admin/Marketing/PaginaInicial/GridController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Admin\Marketing\PaginaInicial;
 
 use App\Enums\Marketing\PaginaInicial\Grid\Formato;
+use App\Enums\Marketing\PaginaInicial\StatusVersao;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Admin\Marketing\PaginaInicial\Grid\AdicionarGridRequest;
 use App\Http\Requests\Admin\Marketing\PaginaInicial\Grid\AdicionarImagemRequest;
@@ -12,6 +13,7 @@ use App\Models\Marketing\PaginaInicial\Grid\GridImagem;
 use App\Models\Marketing\PaginaInicial\Versao;
 use App\Services\Site\PaginaInicialService;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Validation\ValidationException;
 use Inertia\Inertia;
 
 class GridController extends Controller
@@ -29,6 +31,10 @@ class GridController extends Controller
 
     public function salvar(AdicionarGridRequest $request, Versao $versao): mixed
     {
+        if ($versao->status !== StatusVersao::Criado) {
+            $nome = $versao->status->name;
+            abort(403, "Não é possivel editar uma versao com status $nome");
+        }
         $grid = Grid::create($request->only('formato'));
 
         $componente = new Componente($request->only([
@@ -51,20 +57,39 @@ class GridController extends Controller
 
         $grid->load([
             'imagens' => fn ($query) => $query->orderBy('ordem'),
+            'componente'
         ]);
 
-        return Inertia::render('Admin/Marketing/PaginaInicial/Grid/Visualizar', compact('versao', 'grid', 'formato'));
+        $alerta = '';
+        $imagensNecessarias = $grid->formato->imagensNecessarias();
+        if ($grid->imagens->count() != $imagensNecessarias) {
+            $alerta = "O Grid deve ter $imagensNecessarias imagens.";
+        }
+
+        return Inertia::render(
+            'Admin/Marketing/PaginaInicial/Grid/Visualizar',
+            compact('versao', 'grid', 'formato', 'alerta')
+        );
     }
 
     public function adicionarImagem(Versao $versao, Grid $grid): mixed
     {
+        if ($versao->status !== StatusVersao::Criado) {
+            $nome = $versao->status->name;
+            abort(403, "Não é possivel editar uma versao com status $nome");
+        }
         return Inertia::render('Admin/Marketing/PaginaInicial/Grid/AdicionarImagem', compact('versao', 'grid'));
     }
 
     public function salvarImagem(AdicionarImagemRequest $request, Versao $versao, Grid $grid): mixed
     {
+        if ($versao->status !== StatusVersao::Criado) {
+            $nome = $versao->status->name;
+            abort(403, "Não é possivel editar uma versao com status $nome");
+        }
         $gridImagem = new GridImagem($request->all());
         $gridImagem->grid_id = $grid->id;
+        $gridImagem->ordem = $grid->imagens->count() + 1;
 
         $imagemDesktop = $request->file('imagem_desktop');
         $imagemDesktop->store(config('equipamentos.path_imagens'));
@@ -81,11 +106,67 @@ class GridController extends Controller
         return redirect()->route('admin.marketing.paginaInicial.layout.grid.visualizar', [$versao, $grid]);
     }
 
+    public function visualizarImagem(Versao $versao, Grid $grid, GridImagem $gridImagem): mixed
+    {
+        return Inertia::render(
+            'Admin/Marketing/PaginaInicial/Grid/VisualizarImagem',
+            compact('versao', 'grid', 'gridImagem')
+        );
+    }
+
     public function excluirImagem(Versao $versao, Grid $grid, GridImagem $gridImagem): mixed
     {
+        if ($versao->status !== StatusVersao::Criado) {
+            $nome = $versao->status->name;
+            abort(403, "Não é possivel editar uma versao com status $nome");
+        }
         Storage::delete(config('equipamentos.path_imagens') . '/' . $gridImagem->nome_desktop);
         Storage::delete(config('equipamentos.path_imagens') . '/' . $gridImagem->nome_mobile);
         $gridImagem->delete();
+
+        return redirect()->route('admin.marketing.paginaInicial.layout.grid.visualizar', [$versao, $grid]);
+    }
+
+    public function ordemAcima(Versao $versao, Grid $grid, GridImagem $gridImagem)
+    {
+        if ($versao->status !== StatusVersao::Criado) {
+            $nome = $versao->status->name;
+            abort(403, "Não é possivel editar uma versao com status $nome");
+        }
+        $ordem = $gridImagem->ordem;
+
+        if ($ordem <= 1) {
+            throw ValidationException::withMessages(['ordem' => 'A imagem já está na primeira posição']);
+        }
+
+        $imagemPosterior = $grid->imagens()->where('ordem', $ordem - 1)->first();
+        $imagemPosterior->ordem = $ordem;
+        $imagemPosterior->save();
+
+        $gridImagem->ordem = $ordem - 1;
+        $gridImagem->save();
+
+        return redirect()->route('admin.marketing.paginaInicial.layout.grid.visualizar', [$versao, $grid]);
+    }
+
+    public function ordemAbaixo(Versao $versao, Grid $grid, GridImagem $gridImagem)
+    {
+        if ($versao->status !== StatusVersao::Criado) {
+            $nome = $versao->status->name;
+            abort(403, "Não é possivel editar uma versao com status $nome");
+        }
+        $ordem = $gridImagem->ordem;
+
+        if ($ordem >= $grid->imagens()->max('ordem')) {
+            throw ValidationException::withMessages(['ordem' => 'A imagem já está na última posição']);
+        }
+
+        $imagemPosterior = $grid->imagens()->where('ordem', $ordem + 1)->first();
+        $imagemPosterior->ordem = $ordem;
+        $imagemPosterior->save();
+
+        $gridImagem->ordem = $ordem + 1;
+        $gridImagem->save();
 
         return redirect()->route('admin.marketing.paginaInicial.layout.grid.visualizar', [$versao, $grid]);
     }
